Adjust profile page
